Refactor posts controllers
The comments routes in both post types had the exact same logic, thus, there were repeated routes in the controllers. To solve this, common routes were put in the abstract Comments controller, and both posts controllers are now child classes of this.

Also, a bug was found and fixed in the Comments repository: the alias used was 'comments' instead of 'comment'.